Added comments for FeedProvider
