+++ server/src/main/java/com/feeder/server/model/GenericData.java
@@ -1,4 +1,4 @@
-package com.feeder.server.models;
+package com.feeder.server.model;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
