+++ server/src/main/java/com/feeder/server/model/GithubData.java
@@ -1,7 +1,7 @@
-package com.feeder.server.models;
+package com.feeder.server.model;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 
 @JsonInclude(Include.NON_EMPTY)
-public class TwitterData extends GenericData {}
+public class GithubData extends GenericData {}
