+++ server/src/main/java/com/feeder/server/model/TwitterData.java
@@ -1,7 +1,7 @@
-package com.feeder.server.models;
+package com.feeder.server.model;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 
 @JsonInclude(Include.NON_EMPTY)
-public class RedditData extends GenericData {}
+public class TwitterData extends GenericData {}
